requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
discord.py==1.3.4
fuzzywuzzy==0.18.0
idna==2.10
markdownify==0.4.1
multidict==4.7.6
numpy==1.19.1
pycparser==2.20
PyNaCl==1.3.0
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
urllib3==1.25.10
websockets==8.1
yarl==1.5.1
